requirements.txt
alembic==1.11.1
blinker==1.6.2
click==8.1.6
cryptocode==0.1
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
mysqlclient==2.2.0
packaging==23.1
pycryptodomex==3.18.0
PyJWT==2.8.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
Werkzeug==2.3.6
zipp==3.16.2
